fix: address PR #224 review feedback
- NotConnected → NotConnectedWrite in all write-requiring handlers
- NewNotConnectedMsg constructor for custom hints (cmd_features.go)
- Timeout returns ValidationError instead of plain fmt.Errorf
- Standardize RawUnitBackend on model.ID
- Registry panic prefix corrected to 'registry:'
- Full qualified names in AlreadyExists/NotFound messages
- Import/export mapping: distinguish not-found from backend errors
- Published REST: propagate backend errors, include qualified names
- Clarify unhandled statement error message in registry dispatch
- Registry.Validate returns ValidationError instead of plain fmt.Errorf
- Split MockBackend into 13 domain-specific files
- Introduce EntityAccessRuleParams struct for AddEntityAccessRule
- Move mpr type extraction TODO to project tracker
