+++ mdl/backend/doc.go
@@ -4,7 +4,7 @@
 // executor from concrete storage (e.g. .mpr files). Each interface
 // groups related read/write operations by domain concept.
 //
-// TODO: Several method signatures currently reference types from sdk/mpr
+// Several method signatures currently reference types from sdk/mpr
 // (e.g. NavigationDocument, FolderInfo, ImageCollection, JsonStructure,
 // JavaAction, EntityMemberAccess, RenameHit). These should eventually be
 // extracted into a shared types package to remove the mpr dependency.
